+++ Broiler.UI/src/Abstractions/Text/Broiler.UI.RichEdit/RichEditCommand.cs
@@ -28,6 +28,8 @@ public enum RichEditCommand
     Italic,
     Underline,
     Strikethrough,
+    AllCaps,
+    SmallCaps,
     SetForeground,
     SetBackground,
     SetFontFamily,